+++ src/palcommon.h
@@ -23,7 +23,7 @@
 
 #include "video.h"
 
-#define PAL_XY(x, y) (unsigned int)(((((unsigned short)(y)) << 16) & 0xFFFF0000) | (((unsigned short)(x)) & 0xFFFF))
+#define PAL_XY(x, y) (unsigned int)(((((unsigned int)(y)) << 16) & 0xFFFF0000) | (((unsigned int)(x)) & 0xFFFF))
 #define PAL_X(xy) (short)((xy) & 0xFFFF)
 #define PAL_Y(xy) (short)(((xy) >> 16) & 0xFFFF)
 #define PAL_XY_OFFSET(xy, x, y) (unsigned int)(((((int)(y) << 16) & 0xFFFF0000) + ((xy) & 0xFFFF0000)) | (((int)(x) & 0xFFFF) + ((xy) & 0xFFFF)))
